feat: Made the createTestApp able to pass the router Schema type
After this commit, the function createTestApp can be used with
the testClient function and the types of the router
will flow through.

First I updated the AppOpenAPI type to accept a generic Schema or
a empty default. It's the same way as the OpenAPIHone type.

Then I changed the createTestApp to accept a schema generic and
pass it to AppOpenAPI type.

Finally in the tasks.test it can now use this function and the types will work.